improved enemies and bullets
