+++ Assets/Resources/Bullet.prefab
@@ -10,11 +10,11 @@ GameObject:
   m_Component:
   - component: {fileID: 3495878312024126307}
   - component: {fileID: 5645937232986539594}
-  - component: {fileID: 7497542938016069831}
   - component: {fileID: 2103855120813950680}
   - component: {fileID: 4506724329831867283}
+  - component: {fileID: -59651961524077219}
   m_Layer: 7
-  m_Name: Farmer_bullet
+  m_Name: Bullet
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -87,19 +87,6 @@ SpriteRenderer:
   m_WasSpriteAssigned: 1
   m_MaskInteraction: 0
   m_SpriteSortPoint: 0
---- !u!114 &7497542938016069831
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1532884713818425708}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: afe70a508c2174e4b842355e0f566a93, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  duration: 5
 --- !u!50 &2103855120813950680
 Rigidbody2D:
   serializedVersion: 4
@@ -162,3 +149,15 @@ CircleCollider2D:
   m_Offset: {x: 0, y: 0}
   serializedVersion: 2
   m_Radius: 0.5
+--- !u!114 &-59651961524077219
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1532884713818425708}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: b1b0dce83c924eca3afce4aa1d09abf1, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
