--- Assets/Resources/Sheriff_bullet.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 0d000fbaabe3523d38d5e5af7233b10d
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
