+++ Assets/Scripts/Enemies/Bullets/Bullet.cs
@@ -1,21 +1,17 @@
 using System.Collections;
 using System.Collections.Generic;
+using System.Text.RegularExpressions;
 using UnityEngine;
 
-public class Asteroid : ParentBullet
+public class Bullet : ParentBullet
 {
     public override void Setup(Vector3 Dir)
     {
         rb = transform.GetComponent<Rigidbody2D>();
+
         float rotation = Mathf.Rad2Deg * Mathf.Atan2(Dir.y, Dir.x);
         gameObject.transform.rotation = Quaternion.AngleAxis(rotation, Vector3.forward);
         rb.velocity = Dir * speed;
-    }
-
-    public void Update() 
-    {
-        float spinSpeed = Mathf.Sqrt(speed / 10.0f);
-        transform.Rotate(0, 0, 100*Time.deltaTime*spinSpeed);
     }
 
     public override void FixedUpdate()
@@ -26,5 +22,5 @@ public class Asteroid : ParentBullet
             Destroy(gameObject);
         }
     }
-}
 
+}
