--- Assets/Scripts/Enemies/Bullets/SheriffBullet.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6e14f339a2eca9a72bc5113a80dc46a3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
