+++ Assets/Scripts/Enemies/Farmer.cs
@@ -9,7 +9,7 @@ public class Farmer : EnemyParent
 {
     public void Awake()
     {
-        bullet = (GameObject)Resources.Load("Farmer_Bullet");
+        bullet = (GameObject)Resources.Load("Bullet");
 
     }
     public override void Shoot()
@@ -23,7 +23,7 @@ public class Farmer : EnemyParent
             offsetVector.y += Random.Range(-0.15f, 0.15f);
 
             GameObject bulletTransform = Instantiate(bullet, bulletSpawn, Quaternion.identity);            
-            bulletTransform.GetComponent<Farmerbullet>().Setup(offsetVector);
+            bulletTransform.GetComponent<Bullet>().Setup(offsetVector);
         }
     }
